Adding () => is to defer the creation of the UI component, this was causing issues with npm run start.
The issue is <DndStatCharacterUIPage> component is created immediately, and the resulting JavaScript object is stored in the application's central state.
In a production build, React freezes this object for performance. This caused a conflict with state management library (Immer),
which also tries to manage the object, leading to the "Cannot assign to read only property" error.

* Removed context7 dependency from package.json as it is not used anywhere in the codebase.
* Removed test plugins (test-plugin and test-ui-plugin) as they are no longer needed.